# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booker API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/boardroom": {
            "get": {
                "tags": [
                    "Boardroom"
                ],
                "summary": "List of all boardrooms",
                "operationId": "792cad9f85f9e7fc920495bdf9512783",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "events",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Boardroom 1"
                                                    },
                                                    "active": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "post": {
                "tags": [
                    "Boardroom"
                ],
                "summary": "Creating a new boardroom.",
                "operationId": "4c3ecb0f008309a474c7fb4d7c4ddb29",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "name": "Boardroom 1",
                                "active": true
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Boardroom 1"
                                                },
                                                "active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/boardroom/{boardroom}": {
            "put": {
                "tags": [
                    "Boardroom"
                ],
                "summary": "Update boardroom",
                "operationId": "429ee725825ed5c6e421088def835cd8",
                "parameters": [
                    {
                        "name": "boardroom",
                        "in": "path",
                        "description": "Id of the boardroom",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Boardroom 2"
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Boardroom 2"
                                                },
                                                "active": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Boardroom"
                ],
                "summary": "Delete the boardroom",
                "operationId": "00d69cb5300986a1f5435bb7a6f288f3",
                "parameters": [
                    {
                        "name": "boardroom",
                        "in": "path",
                        "description": "Id of the boardroom",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}